5917 TOGGLE Ikke avrund avgiftspliktigInntektMnd mot frontend (#1839)
Verdien er allerede avrundet ved beregning, så dette er både unødvendig og kan risikere å vise feil verdi dersom noe har skjedd galt.

+++ frontend-api/src/main/kotlin/no/nav/melosys/tjenester/gui/dto/trygdeavgift/InntekskildeDto.kt
@@ -3,16 +3,17 @@ package no.nav.melosys.tjenester.gui.dto.trygdeavgift
 import no.nav.melosys.domain.avgift.Inntektsperiode
 import no.nav.melosys.domain.kodeverk.Inntektskildetype
 import no.nav.melosys.service.avgift.dto.InntektskildeRequest
+import java.math.BigDecimal
 
 data class InntekskildeDto(
     val type: Inntektskildetype,
     val arbeidsgiversavgiftBetales: Boolean,
-    val avgiftspliktigInntektMnd: Int
+    val avgiftspliktigInntektMnd: BigDecimal
 ) {
     constructor(inntektsperiode: Inntektsperiode) : this(
         inntektsperiode.type,
         inntektsperiode.isArbeidsgiversavgiftBetalesTilSkatt,
-        inntektsperiode.avgiftspliktigInntektMnd.verdi.toInt()
+        inntektsperiode.avgiftspliktigInntektMnd.verdi
     )
 
     fun tilRequest(): InntektskildeRequest =
